requirements.txt
# ===============================
# FastAPI and ASGI server
# ===============================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ===============================
# Database and caching
# ===============================
chromadb==0.4.15
motor>=3.6,<4
pymongo>=4.9,<5
redis>=4.6.0,<5.0.0
aiocache==0.12.2

# ===============================
# Files & images
# ===============================
Pillow==10.1.0
aiofiles==23.2.1

# ===============================
# Auth & security
# ===============================
python-jose[cryptography]==3.3.0
PyJWT>=2.8,<3
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# ===============================
# HTTP clients & utils
# ===============================
httpx==0.25.2
aiohttp==3.9.1

# ===============================
# Validation & settings
# ===============================
pydantic>=2.7.4,<3.0.0
pydantic-settings==2.6.1
email-validator==2.1.0
setuptools>=68.0.0
numpy>=1.26.0
posthog<3   # for chromadb 0.4.15 telemetry compatibility

# ===============================
# LangChain ecosystem (aligned pins; do NOT pin langchain-core separately)
# ===============================
langchain==0.3.10
langchain-openai==0.2.12          # keep if you use OpenAI for chat; safe with LC 0.3.10
langchain-community==0.3.10
langchain-chroma==0.2.2
langchain-text-splitters==0.3.4

# ===============================
# Local embeddings via Hugging Face
# (pulls transformers + tokenizers through sentence-transformers)
# ===============================
sentence-transformers==2.6.1
# Optional explicit pins for stability; uncomment if you see resolver churn:
# transformers==4.45.2
# tokenizers==0.19.1
# huggingface-hub==0.25.2
# accelerate==0.34.2

# ===============================
# Document loaders / RAG helpers
# (may need apt packages: poppler-utils, tesseract-ocr, libgl1, etc.)
# ===============================
pypdf>=4.0.0
python-docx>=1.0.0
pytesseract==0.3.10
pdf2image==1.16.3
pypdfium2>=4.20.0
unstructured>=0.11.0

# ===============================
# Rate limiting
# ===============================
slowapi==0.1.9
limits==3.6.0

# ===============================
# Background tasks
# ===============================
celery[redis]==5.3.4

# ===============================
# Optional: OpenAI client (NOT used for embeddings here).
# Leave commented if not needed anywhere else in your app.
# ===============================
openai
